Create a separate util module for all non-public utilities
