+++ src/util/mod.rs
@@ -1,0 +1,3 @@
+//! Various random non-public utilities used within the crate.
+
+pub(crate) mod cow;
